requirements.txt
streamlit
PyPDF2
pdf2image
Pillow
reportlab